refactor(install): split install.ts into per-task factories under install-tasks/
Extracts each inline task closure from the 1.5k-line install.ts into its own
factory file under src/cli/commands/install-tasks/, keeping install.ts as a
thin entry point that just builds the context and runs buildInstallTasks().

- context.ts owns the shared InstallCtx / InstallOptions / GetRepoSnapshotFn
  types so task files don't have to reach back into install.ts.
- helpers/ holds the non-task utilities (git checks, repo snapshot wrapper,
  browser opener, skill discovery, installOneSkill, cleanupRemovedSkills,
  tool-warning collectors, required-command check).
- One file per task (verify-prerequisites, resolve-plugins, install-skills,
  configure-tools, …), each exporting a factory returning Task<InstallCtx>.
- index.ts re-exports the context types and assembles the ordered task list.

install.ts: 1525 -> 140 lines. No behavior change; all 780 tests pass.
