+++ src/cli/commands/install-tasks/helpers/text.ts
@@ -1,0 +1,6 @@
+export function indentLines(text: string, prefix: string): string {
+  return text
+    .split('\n')
+    .map((line) => (line.length > 0 ? `${prefix}${line}` : line))
+    .join('\n');
+}
